move git relative code to MBO\RemoteGit
