+++ src/MBO/RemoteGit/GithubClient.php
@@ -1,10 +1,10 @@
 <?php
 
-namespace MBO\SatisGitlab\Git;
+namespace MBO\RemoteGit;
 
 use \GuzzleHttp\Client as GuzzleHttpClient;
 use Psr\Log\LoggerInterface;
-use MBO\SatisGitlab\Filter\ProjectFilterInterface;
+use MBO\RemoteGit\Filter\ProjectFilterInterface;
 
 /**
  * Client implementation for github
