+++ tests/Git/GitlabClientTest.php
@@ -5,12 +5,12 @@ namespace Tests\SatisGitlab\Git;
 use Tests\SatisGitlab\TestCase;
 
 use GuzzleHttp\Client as GuzzleHttpClient;
-use MBO\SatisGitlab\Git\GitlabClient;
+use MBO\RemoteGit\GitlabClient;
 
 use Psr\Log\NullLogger;
-use MBO\SatisGitlab\Git\ClientOptions;
-use MBO\SatisGitlab\Git\ClientFactory;
-use MBO\SatisGitlab\Git\FindOptions;
+use MBO\RemoteGit\ClientOptions;
+use MBO\RemoteGit\ClientFactory;
+use MBO\RemoteGit\FindOptions;
 
 
 class GitlabClientTest extends TestCase {
