+++ tests/TestCase.php
@@ -4,7 +4,7 @@ namespace Tests\SatisGitlab;
 
 use PHPUnit\Framework\TestCase as BaseTestCase;
 
-use MBO\SatisGitlab\Git\ProjectInterface;
+use MBO\RemoteGit\ProjectInterface;
 
 class TestCase extends BaseTestCase {
